refactor: relocate and rename analyzeview to assemblyview in the shared package (#1604)
