+++ packages/shared/views/AssemblyView/components/Main/main.styles.ts
