+++ packages/shared/views/AssemblyView/components/Top/top.tsx
@@ -4,10 +4,10 @@ import { type Props } from "./types";
 import { getBreadcrumbs } from "./utils";
 
 /**
- * Top component for the AnalyzeView, which displays the page title and breadcrumbs.
+ * Top component for the assembly detail view, which displays the page title and breadcrumbs.
  * @param props - Component props.
  * @param props.assembly - Assembly.
- * @returns A JSX element representing the top section of the AnalyzeView.
+ * @returns A JSX element representing the top section of the assembly detail view.
  */
 export const Top = ({ assembly }: Props): JSX.Element => {
   return (
