+++ packages/shared/views/AssemblyView/components/Top/types.ts
@@ -1,0 +1,5 @@
+import { type AssemblyContract } from "@repo/shared/apis/types";
+
+export interface Props {
+  assembly: AssemblyContract;
+}
